anomaly_detector/requirements.txt
connexion[flask,uvicorn,swagger-ui]
httpx
pykafka
flask-cors
PyYAML
setuptools
